+++ Makefile
@@ -3,6 +3,7 @@ PROJ_DIR := $(dir $(abspath $(lastword $(MAKEFILE_LIST))))
 # Configuration of extension
 EXT_NAME=rpc
 EXT_CONFIG=${PROJ_DIR}extension_config.cmake
+EXT_FLAGS=-DCMAKE_CXX_STANDARD=17
 
 # Include the Makefile from extension-ci-tools
 include extension-ci-tools/makefiles/duckdb_extension.Makefile
