+++ src/include/quack_message.hpp
@@ -1,8 +1,6 @@
 #pragma once
 
-#include "duckdb/common/serializer/binary_deserializer.hpp"
 #include "duckdb/common/serializer/binary_serializer.hpp"
-#include "duckdb/common/serializer/memory_stream.hpp"
 #include "duckdb/parser/parsed_data/drop_info.hpp"
 
 namespace duckdb {
