+++ src/include/quack_startstop.hpp
