--- src/include/rpc_scan_function.hpp
@@ -1,16 +1,0 @@
-#pragma once
-
-namespace duckdb {
-
-class TableFunction;
-
-class RpcScanFunction {
-public:
-	static TableFunction GetFunction();
-};
-
-class RpcScanByNameFunction {
-public:
-	static TableFunction GetFunction();
-};
-} // namespace duckdb
